#32: Move method to test utils.
