+++ src/test/java/com/commercetools/sync/commons/MockUtils.java
@@ -1,12 +1,16 @@
 package com.commercetools.sync.commons;
 
+import com.commercetools.sync.commons.helpers.BaseSyncStatistics;
 import com.commercetools.sync.commons.helpers.CtpClient;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
 import io.sphere.sdk.client.BlockingSphereClient;
 import io.sphere.sdk.client.SphereClientConfig;
 import io.sphere.sdk.types.CustomFieldsDraft;
 
+import javax.annotation.Nonnull;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -48,5 +52,19 @@ public class MockUtils {
         when(ctpClient.getClientConfig()).thenReturn(SphereClientConfig.of(CTP_KEY, CTP_ID, CTP_SECRET));
         when(ctpClient.getClient()).thenReturn(client);
         return ctpClient;
+    }
+
+    /**
+     * Builds a JSON String that represents the fields of the supplied instance of {@link BaseSyncStatistics}.
+     * Note: The order of the fields in the built JSON String depends on the order of the instance variables in this
+     * class.
+     *
+     * @param statistics the instance of {@link BaseSyncStatistics} from which to create a JSON String.
+     * @return a JSON representation of the given {@code statistics} as a String.
+     */
+    public static String getStatisticsAsJsonString(@Nonnull final BaseSyncStatistics statistics)
+        throws JsonProcessingException {
+        final ObjectMapper mapper = new ObjectMapper();
+        return mapper.writeValueAsString(statistics);
     }
 }
